Improve listing with invalid oracles (#620)
* Stable Price: Reset on first price != 0

This helps when listing tokens or perp markets with an upcoming
oracle. Currently the stable price would be 0 and would need to be
manually reset by DAO proposal.

With this change, the first non-zero value will be used as the starting
point for the stable price instead.

* Listing instructions: Succeed even if oracle is invalid

But require the stable price to reset once it becomes valid.